Development

+++ app/jobs/update_etablissement_rows_job.rb
@@ -1,33 +1,26 @@
-class UpdateEtablissementRowsJob
+class UpdateEtablissementRowsJob < EtablissementRowJobs
   attr_accessor :lines
   def initialize(lines)
     @lines = lines
   end
 
   def perform
-    etablissements = []
+    etablissements = fill_etablissements
 
-    for line in lines do
-      etablissement_atts = EtablissementAttrsFromLine.instance.call(line)
-      etablissements << etablissement_atts
-    end
+    update_attrs(etablissements)
+  end
 
+  def update_attrs(etablissements)
     etablissements.each do |etablissement_attrs|
-      siret = etablissement_attrs[:siret]
-      etablissement = Etablissement.find_or_initialize_by(siret: siret)
+      etablissement = Etablissement.find_or_initialize_by(siret: etablissement_attrs[:siret])
 
       nature_mise_a_jour = etablissement_attrs[:nature_mise_a_jour]
-
-      if nature_mise_a_jour == 'I'
-        # Il y a une paire I/F, I étant l'état initial
-        next
-      elsif nature_mise_a_jour == 'E'
-        # On supprime l'établissement si il est persisté
-        # Si non persisté veut dire qu on rejoue un patch interrompu
-        if etablissement.persisted?
-          etablissement.destroy
-        end
-
+      # Il y a une paire I/F, I etant l'etat initial
+      next if nature_mise_a_jour == 'I'
+      if nature_mise_a_jour == 'E'
+        # On supprime l'etablissement si il est persiste
+        # Si non persiste veut dire qu on rejoue un patch interrompu
+        etablissement.destroy if etablissement.persisted?
         next
       end
       etablissement.update_attributes(etablissement_attrs)
